no need for api_fixtures separate module
add a few conditions to user tests
